Solved!

+++ assignments/08Caracol/src/exercise.py
@@ -2,7 +2,7 @@ def main():
     #escribe tu código abajo de esta línea
     minutos= float(input("Dame los minutos: "))
     centimetros= (60*5.7) / 10
-    total= centimetros*minutos
+    total= float(centimetros*minutos)
     
     print("Centímetros recorridos:", total)
